Fix generation of the documentation.
We use a manual list of folder now. To support different archtecture in
the different repo.

Next step is to support custom output path. Like samples/pixi/basic. For
the moment it's support only samples/basic. But we need more control on
the out directory.

Remove basic.js file

Samples now embedded

Switch to use a manual list of folder.